refactor: 서비스 계층 분리 및 JPA Repository를 인프라 계층으로 이전 #29
