+++ caregiver-service/src/main/java/com/carenest/business/caregiverservice/domain/service/CaregiverDomainService.java
@@ -1,0 +1,4 @@
+package com.carenest.business.caregiverservice.domain.service;
+
+public interface CaregiverDomainService {
+}
