+++ caregiver-service/src/main/java/com/carenest/business/caregiverservice/domain/service/CaregiverDomainServiceImpl.java
@@ -1,0 +1,4 @@
+package com.carenest.business.caregiverservice.domain.service;
+
+public class CaregiverDomainServiceImpl {
+}
