--- caregiver-service/src/main/java/com/carenest/business/caregiverservice/infrastructure/repository/CaregiverCustomRepository.java
@@ -1,4 +1,0 @@
-package com.carenest.business.caregiverservice.infrastructure.repository;
-
-public interface CaregiverCustomRepository {
-}
