+++ caregiver-service/src/main/java/com/carenest/business/caregiverservice/infrastructure/repository/querydsl/CaregiverCustomRepository.java
@@ -1,0 +1,4 @@
+package com.carenest.business.caregiverservice.infrastructure.repository.querydsl;
+
+public interface CaregiverCustomRepository {
+}
