+++ caregiver-service/src/main/java/com/carenest/business/caregiverservice/infrastructure/repository/querydsl/CaregiverRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.carenest.business.caregiverservice.infrastructure.repository;
+package com.carenest.business.caregiverservice.infrastructure.repository.querydsl;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
