+++ caregiver-service/src/main/java/com/carenest/business/caregiverservice/presentation/controller/CaregiverController.java
@@ -16,7 +16,7 @@ import com.carenest.business.caregiverservice.application.dto.request.CaregiverC
 import com.carenest.business.caregiverservice.application.dto.response.CaregiverCreateResponseServiceDTO;
 import com.carenest.business.caregiverservice.application.dto.response.CaregiverReadResponseServiceDTO;
 import com.carenest.business.caregiverservice.application.dto.response.CaregiverUpdateResponseServiceDTO;
-import com.carenest.business.caregiverservice.application.service.CaregiverService;
+import com.carenest.business.caregiverservice.application.service.CaregiverServiceImpl;
 import com.carenest.business.caregiverservice.presentation.dto.request.CaregiverUpdateRequestDTO;
 import com.carenest.business.caregiverservice.presentation.dto.response.CaregiverReadResponseDTO;
 import com.carenest.business.caregiverservice.presentation.dto.mapper.CaregiverPresentationMapper;
@@ -34,7 +34,7 @@ import lombok.extern.slf4j.Slf4j;
 @RequestMapping("/api/v1/caregivers")
 public class CaregiverController {
 
-	private final CaregiverService caregiverService;
+	private final CaregiverServiceImpl caregiverServiceImpl;
 
 	@Qualifier("caregiverPresentationMapper")
 	private final CaregiverPresentationMapper presentationMapper;
@@ -45,14 +45,14 @@ public class CaregiverController {
 		@RequestBody CaregiverCreateRequestDTO createRequestDTO
 	){
 		CaregiverCreateRequestServiceDTO requestServiceDTO = presentationMapper.toCreateServiceDto(createRequestDTO);
-		CaregiverCreateResponseServiceDTO responseDTO = caregiverService.createCaregiver(requestServiceDTO);
+		CaregiverCreateResponseServiceDTO responseDTO = caregiverServiceImpl.createCaregiver(requestServiceDTO);
 		return ResponseDto.success("서비스 등록 요청이 접수되었습니다. 관리자 승인 후 활성화됩니다.",presentationMapper.toCreateResponseDto(responseDTO));
 	}
 
 	// Read (개인 조회)
 	@GetMapping("/{caregiverId}")
 	public ResponseDto<CaregiverReadResponseDTO> getCaregiverDetail(@PathVariable UUID caregiverId){
-		CaregiverReadResponseServiceDTO responseDTO = caregiverService.getCaregiver(caregiverId);
+		CaregiverReadResponseServiceDTO responseDTO = caregiverServiceImpl.getCaregiver(caregiverId);
 		return ResponseDto.success(presentationMapper.toReadResponseDto(responseDTO));
 	}
 
@@ -62,7 +62,7 @@ public class CaregiverController {
 		@PathVariable UUID caregiverId,
 		@RequestBody CaregiverUpdateRequestDTO requestDTO
 	){
-		CaregiverUpdateResponseServiceDTO responseDTO = caregiverService.updateCaregiver(caregiverId,requestDTO);
+		CaregiverUpdateResponseServiceDTO responseDTO = caregiverServiceImpl.updateCaregiver(caregiverId,requestDTO);
 		return ResponseDto.success("간병인 정보가 수정되었습니다.",presentationMapper.toUpdateResponseDto(responseDTO));
 	}
 
@@ -71,7 +71,7 @@ public class CaregiverController {
 	public ResponseDto<Void> deleteCaregiver(
 		@PathVariable UUID caregiverId
 	){
-		caregiverService.deleteCaregiver(caregiverId);
+		caregiverServiceImpl.deleteCaregiver(caregiverId);
 		return ResponseDto.success("간병인 정보가 삭제되었습니다.",null);
 	}
